Memory usage optimization

+++ terra-core/src/main/java/com/ritualsoftheold/terra/core/material/TerraObject.java
@@ -14,25 +14,25 @@ public class TerraObject {
         return new Builder();
     }
 
-    private int x;
-    private int y;
-    private int z;
+    private float x;
+    private float y;
+    private float z;
 
-    public void position(int x, int y, int z){
+    public void position(float x, float y, float z){
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public int getZ() {
+    public float getZ() {
         return z;
     }
